// src/types/coin.ts
// src/types/coin.ts

export interface Coin {
  id: string;
  symbol: string;
  name: string;
  image: string;
  current_price: number;
  price_change_percentage_24h: number;
  market_cap: number;
  market_cap_rank: number;
}

export interface CoinDetail extends Omit<Coin, 'image'> {
  image: {
    thumb: string;
    small: string;
    large: string;
  };
  description: { en: string };
  market_data: {
    current_price: { usd: number };
    price_change_percentage_24h: number;
    price_change_percentage_7d: number;
    market_cap: { usd: number };
    total_volume: { usd: number };
  };
}

export interface ChartData {
  date: string;
  price: number;
}